Use "clone" method for bilinear resampling
When drawing a bilinear-scaled bitmap with agg, we used to interpolate
the edge pixels assuming everything out of the picture was black. This
led to an ugly dark border around the scaled image.

Instead, use the "clone" method, which repeats the last line or column
of pixels around the image, leading to more correct results in most
cases.

This also helps with some of WebKit tests, which are trying to use
images as references and don't expect an extra border around them.