Icon-O-Matic: Forget fMessageAfterSave when FilePanel was cancelled.
fMessageAfterSave capture what was to be done after a file panel was
used. For example, picking up quitting the app after the user saved
unsaved changes. The bug was that when trying to quit IOM, deciding to
save, but then cancelling the file panel, the quitting stopped (as expected).
But then the next time the document was saved, IOM quit unexpectedly.