WebPositive: BTextView shortcuts work in URL bar
Pass Cmd+Left/Right to the BTextView base class if the URL bar is focused.
This allows us to make use of the navigation shortcuts in BTextView when we
are focused on the URL bar, and activates back and forward navigation
otherwise.

Unfortunately this doesn't also work for text fields inside webpages.

Also make modifierKeys signed.