Don't reuse same tab for multiple bookmarks.
As WebKit is asynchronous to the window, when launching a request in a
BWebView, IsBlankTab() will keep returning true until it gets the
BMessage and updates its state.

When opening bookmarks or refs, we would send them too fast, not detect
this, and reuse the same tab for several items. Make sure the blank tab
is only used once when looping over the refs, and force opening all
remaining refs in new tabs of the same window.

Fixes #6625.

Also optimizes the ref loading by not looking up the window for each
ref. Pick one window, then use it for all the bookmarks in the loop.