Remove FIND_NEXT message for find text control
BTextControl calls Invoke() when it loses the focus and the text has
changed. We don't want this to trigger a search, and instead rely on the
user pressing enter to run the search.

Fixes the search happening again when closing the search bar (escape key
or otherwise), or clicking on the page after searching. Fixes #10168.