Icons Screensaver: Get icons from MIME db again
The previous method only queried application icons, not document icons.

Unfortunately, the app icons in the MIME DB are broken/not present.
IconsSaver only displays document icons for the time being.

Because of better error checking the garbled icons are filtered out though
at least.

Next step is to fix the application icons in the MIME DB.

Some other changes:
* Don't draw if the BBitmap was not filled out correctly
* Add Vincent Duvert to authors list, he's already in copyright
* Convert fVectorIcons from a BList to a BObjectList
* Put vector_icon struct in it’s own header (needed for above)
* Remove type param from vector_icon struct
* Bump max icon count to 300, hopefully this should be enough, 128
  is too few for app and document icons in default install.