ServerWindow: Fix converting the clip picture origin
The current view state is already used for the picture playback, the
remaining conversion needs to take into account screen location and scrolling
offset of the view. This may need further investigation: a) When is this
updated? It needs to be whenever a View changes screen location, size or scroll
offset, which may not be the times when this method is called. b) I am not
sure if the scrolling offset is indeed part of this conversion...