exfat: style fix param/variable renaming
Pass the partition size into the default name, not the device size,
this makes the variable names match the comment.

There is no functional change here, the variable names got renamed
to make it more clear what's happening. We want show the partition
size, not the device size, and we want to error if the device size is less
than the partition size, not if the partition size is less than the device size.