PackageInstaller: Install apps to non-packaged "apps" folder.
Previously, the target folder was always unset, which meant one needed
to select a target folder manually. When running from a read-only media,
this would still be a problem (a TODO), so that instead of failing to install
the package, the user is simply prompted for a path.
The change means that together with rewriting old paths to packaged
locations to the non-packaged equivalents, just double clicking a PKG and
hitting "Install" will now work. Tested with Moho installer and
Gobe Productive.