Replace a few equ with %define to make nasm happy
Interestingly, it only complained about some of them. Changing them all
to %define though, as equ means it should try to evaluate it once, which
of course it can't really if those are ebp+CONSTANT.

Didn't catch those since I wasn't building gcc2h.
Thanks to jessicah for reporting and testing the fix!